Extract terminal log PTY module (oh-tab-vzgb)
